// lib/audio/types.ts
export type TTSProviderId =
  | 'openai-tts'
  | 'elevenlabs-tts'
  | 'browser-native-tts'
  | 'puter-tts'
  | 'azure-tts'
  | 'glm-tts'
  | 'qwen-tts'
  | 'minimax-tts'
  | 'doubao-tts';